refactor: 설문조사1 컴포넌트로 수정 및 상수 분리
